Ported everything to NS2.0 and rebranded the package to eshopworld
