--- src/DevOpsFlex.Messaging.sln
@@ -1,47 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27428.2037
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DevOpsFlex.Messaging", "DevOpsFlex.Messaging\DevOpsFlex.Messaging.csproj", "{70BC70A5-C623-480E-BB68-982C46011441}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A7CF6C38-1E42-48AC-A686-9002B1C5C47F}"
-	ProjectSection(SolutionItems) = preProject
-		codecoverage.runsettings = codecoverage.runsettings
-		..\README.md = ..\README.md
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DevOpsFlex.Messaging.Tests", "Tests\DevOpsFlex.Messaging.Tests\DevOpsFlex.Messaging.Tests.csproj", "{966B329C-E89E-44C2-AFD0-476944307483}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{70BC70A5-C623-480E-BB68-982C46011441}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{70BC70A5-C623-480E-BB68-982C46011441}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{70BC70A5-C623-480E-BB68-982C46011441}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{70BC70A5-C623-480E-BB68-982C46011441}.Debug|x86.Build.0 = Debug|Any CPU
-		{70BC70A5-C623-480E-BB68-982C46011441}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{70BC70A5-C623-480E-BB68-982C46011441}.Release|Any CPU.Build.0 = Release|Any CPU
-		{70BC70A5-C623-480E-BB68-982C46011441}.Release|x86.ActiveCfg = Release|Any CPU
-		{70BC70A5-C623-480E-BB68-982C46011441}.Release|x86.Build.0 = Release|Any CPU
-		{966B329C-E89E-44C2-AFD0-476944307483}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{966B329C-E89E-44C2-AFD0-476944307483}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{966B329C-E89E-44C2-AFD0-476944307483}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{966B329C-E89E-44C2-AFD0-476944307483}.Debug|x86.Build.0 = Debug|Any CPU
-		{966B329C-E89E-44C2-AFD0-476944307483}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{966B329C-E89E-44C2-AFD0-476944307483}.Release|Any CPU.Build.0 = Release|Any CPU
-		{966B329C-E89E-44C2-AFD0-476944307483}.Release|x86.ActiveCfg = Release|Any CPU
-		{966B329C-E89E-44C2-AFD0-476944307483}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {0486A538-0099-4EEC-B48D-177FAB22B6AE}
-	EndGlobalSection
-EndGlobal
