--- src/DevOpsFlex.Messaging/packages.config
@@ -1,61 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Eshopworld.Core" version="1.0.2" targetFramework="net462" />
-  <package id="JetBrains.Annotations" version="11.1.0" targetFramework="net462" />
-  <package id="Microsoft.Azure.Amqp" version="2.3.0" targetFramework="net462" />
-  <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.AppService.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Batch.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.BatchAI.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Cdn.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Compute.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.ContainerInstance.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.ContainerRegistry.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.ContainerService.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.CosmosDB.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Dns.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.EventHub.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Graph.RBAC.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.KeyVault.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Locks.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Monitor.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Msi.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Network.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Redis.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.ResourceManager.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Search.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.ServiceBus.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Sql.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.Storage.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.Management.TrafficManager.Fluent" version="1.11.1" targetFramework="net462" />
-  <package id="Microsoft.Azure.ServiceBus" version="3.0.2" targetFramework="net462" />
-  <package id="Microsoft.Azure.Services.AppAuthentication" version="1.0.1" targetFramework="net462" />
-  <package id="Microsoft.Data.Edm" version="5.8.2" targetFramework="net462" />
-  <package id="Microsoft.Data.OData" version="5.8.2" targetFramework="net462" />
-  <package id="Microsoft.Data.Services.Client" version="5.8.2" targetFramework="net462" />
-  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="3.17.2" targetFramework="net462" />
-  <package id="Microsoft.IdentityModel.Logging" version="5.2.2" targetFramework="net462" />
-  <package id="Microsoft.IdentityModel.Tokens" version="5.2.2" targetFramework="net462" />
-  <package id="Microsoft.Rest.ClientRuntime" version="2.3.9" targetFramework="net462" />
-  <package id="Microsoft.Rest.ClientRuntime.Azure" version="3.3.10" targetFramework="net462" />
-  <package id="Microsoft.Rest.ClientRuntime.Azure.Authentication" version="2.3.2" targetFramework="net462" />
-  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net462" />
-  <package id="System.ComponentModel.EventBasedAsync" version="4.0.11" targetFramework="net462" />
-  <package id="System.Diagnostics.DiagnosticSource" version="4.4.1" targetFramework="net462" />
-  <package id="System.Dynamic.Runtime" version="4.0.0" targetFramework="net462" />
-  <package id="System.IdentityModel.Tokens.Jwt" version="5.2.2" targetFramework="net462" />
-  <package id="System.Linq.Queryable" version="4.0.0" targetFramework="net462" />
-  <package id="System.Net.Requests" version="4.0.11" targetFramework="net462" />
-  <package id="System.Net.WebSockets" version="4.0.0" targetFramework="net462" />
-  <package id="System.Net.WebSockets.Client" version="4.0.2" targetFramework="net462" />
-  <package id="System.Reactive" version="4.0.0" targetFramework="net462" />
-  <package id="System.Runtime.Serialization.Primitives" version="4.3.0" targetFramework="net462" />
-  <package id="System.Runtime.Serialization.Xml" version="4.3.0" targetFramework="net462" />
-  <package id="System.Security.Cryptography.Algorithms" version="4.2.0" targetFramework="net462" />
-  <package id="System.Security.Cryptography.Encoding" version="4.0.0" targetFramework="net462" />
-  <package id="System.Security.Cryptography.Primitives" version="4.0.0" targetFramework="net462" />
-  <package id="System.Security.Cryptography.X509Certificates" version="4.1.0" targetFramework="net462" />
-  <package id="System.Spatial" version="5.8.2" targetFramework="net462" />
-  <package id="WindowsAzure.Storage" version="9.2.0" targetFramework="net462" />
-</packages>
