+++ src/Eshopworld.Messaging.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27703.2035
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Eshopworld.Messaging", "Eshopworld.Messaging\Eshopworld.Messaging.csproj", "{636592BC-C306-4344-AA04-DDA552DD6DB0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eshopworld.Messaging.Tests", "Tests\Eshopworld.Messaging.Tests\Eshopworld.Messaging.Tests.csproj", "{59440B18-9844-40D5-84DB-0E6762FEB854}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{636592BC-C306-4344-AA04-DDA552DD6DB0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{636592BC-C306-4344-AA04-DDA552DD6DB0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{636592BC-C306-4344-AA04-DDA552DD6DB0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{636592BC-C306-4344-AA04-DDA552DD6DB0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{59440B18-9844-40D5-84DB-0E6762FEB854}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{59440B18-9844-40D5-84DB-0E6762FEB854}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{59440B18-9844-40D5-84DB-0E6762FEB854}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{59440B18-9844-40D5-84DB-0E6762FEB854}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2F2A4416-E4FC-4BC5-A9A0-E49B39139450}
+	EndGlobalSection
+EndGlobal
