+++ src/Eshopworld.Messaging/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Eshopworld.Messaging.Tests")]
+[assembly: InternalsVisibleTo("Eshopworld.Messaging.Tests.Profiler")]
