--- src/Tests/DevOpsFlex.Messaging.Tests.Profiler/packages.config
@@ -1,24 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="4.3.0" targetFramework="net462" />
-  <package id="DevOpsFlex.Tests.Core.Analyzers" version="1.1.0" targetFramework="net462" />
-  <package id="Eshopworld.Tests.Core" version="1.0.1" targetFramework="net462" />
-  <package id="Eshopworld.Tests.Core.Analyzers" version="1.0.0" targetFramework="net462" />
-  <package id="FluentAssertions" version="5.4.0" targetFramework="net462" />
-  <package id="JetBrains.dotMemoryUnit" version="3.0.20171219.105559" targetFramework="net462" />
-  <package id="Microsoft.CodeCoverage" version="1.0.3" targetFramework="net462" />
-  <package id="Microsoft.NET.Test.Sdk" version="15.7.2" targetFramework="net462" />
-  <package id="Microsoft.TestPlatform.TestHost" version="15.7.0" targetFramework="net462" />
-  <package id="Moq" version="4.8.3" targetFramework="net462" />
-  <package id="System.Reactive" version="4.0.0" targetFramework="net462" />
-  <package id="System.Threading.Tasks.Extensions" version="4.3.0" targetFramework="net462" />
-  <package id="System.ValueTuple" version="4.4.0" targetFramework="net462" />
-  <package id="xunit" version="2.3.1" targetFramework="net462" />
-  <package id="xunit.abstractions" version="2.0.1" targetFramework="net462" />
-  <package id="xunit.analyzers" version="0.7.0" targetFramework="net462" />
-  <package id="xunit.assert" version="2.3.1" targetFramework="net462" />
-  <package id="xunit.core" version="2.3.1" targetFramework="net462" />
-  <package id="xunit.extensibility.core" version="2.3.1" targetFramework="net462" />
-  <package id="xunit.extensibility.execution" version="2.3.1" targetFramework="net462" />
-  <package id="xunit.runner.visualstudio" version="2.3.1" targetFramework="net462" developmentDependency="true" />
-</packages>
